requirements.txt

python-telegram-bot==20.6
selenium==4.16.0
python-dotenv==1.0.1
